requirements.txt
django==4.2.7
django-autoslug==1.9.9
pillow==10.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9